10_lesson/requirements.txt
# Основные зависимости проекта
pytest==9.0.1
allure-pytest==2.15.2
allure-python-commons==2.15.2
selenium==4.38.0